requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
greenlet==3.3.2
idna==3.11
iniconfig==2.3.0
packaging==26.0
playwright==1.58.0
pluggy==1.6.0
pyee==13.0.1
Pygments==2.19.2
pytest==9.0.2
pytest-base-url==2.1.0
pytest-playwright==0.7.2
python-slugify==8.0.4
requests==2.32.5
text-unidecode==1.3
typing_extensions==4.15.0
urllib3==2.6.3
